zipcode works city with enum star create but don't have logic yet and I still have to make class controller address create and get all

+++ src/main/java/example/addressrepo/dto/ZipCodeDto.java
@@ -10,7 +10,7 @@ import lombok.Data;
 @Data
 public class ZipCodeDto {
 
-    private Long zipCodeId;
+    private Integer zipCodeId;
 
     @JsonProperty("code")
     @Pattern(regexp = "^\\d{5}(-\\d{4})?$", message = "Invalid zip code format")
